restructure file output
