+++ src/fileio/fiberio.jl
@@ -1,0 +1,2 @@
+function fbrwrite end
+function fbrread end
